Add report level and partner handling
The commit includes the addition of report level handling and partner id parameters in row configuration and python scripts. This enables users to fetch data sorted by either Partner or Advertiser level. The 'partners' field has been introduced to download reports for selected partners similar to the 'advertisers' field. The changes related to this are made in component configuration and the code files - ttd.py and main.py.